project

+++ prepare/Learning03/Fraction.cs
@@ -51,7 +51,7 @@ public class Fraction
 
     public double GetDecimalValue()
     {
-        return (double)_top/_bottom;
+        return (double)_top / (double)_bottom;
     }
